backend/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
awsebcli==3.20.10
bidict==0.23.1
blinker==1.8.2
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
env==0.1.0
eventlet==0.36.1
exceptiongroup==1.2.2
fastapi
Flask==3.0.3
Flask-SocketIO==5.3.6
fonttools==4.52.4
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.2.11
langchain-core==0.2.26
langchain-ollama==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.95
llvmlite==0.39.1
MarkupSafe==2.1.5
matplotlib==3.7.2
multidict==6.0.5
neat-python==0.92
numpy==1.23.1
ollama==0.3.1
openai==1.37.1
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.0
pandas==1.5.3
pathspec==0.10.1
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.9
pycurl==7.45.3
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
PyJWT==2.8.0
pyparsing==3.0.9
pypiwin32==223
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.2
scikit-learn==1.1.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.8.5
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.5.0
termcolor==1.1.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.19
uvicorn
wcwidth==0.1.9
Werkzeug==3.0.3
whitenoise==6.7.0
wincertstore==0.2
wsproto==1.2.0
yarl==1.9.4
